requirements.txt
discord.py==2.3.1
mcstatus==11.0.0
python-dotenv==1.0.0
